#212 tidying up after websockets refactoring
