+++ src/Ocelot/ServiceDiscovery/Errors/UnableToFindServiceDiscoveryProviderError.cs
@@ -1,0 +1,12 @@
+using Ocelot.Errors;
+
+namespace Ocelot.ServiceDiscovery.Errors
+{
+    public class UnableToFindServiceDiscoveryProviderError : Error
+    {
+        public UnableToFindServiceDiscoveryProviderError(string message) 
+            : base(message, OcelotErrorCode.UnableToFindServiceDiscoveryProviderError)
+        {
+        }
+    }
+}
