--- src/Ocelot/ServiceDiscovery/IServiceDiscoveryProvider.cs
@@ -1,11 +1,0 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using Ocelot.Values;
-
-namespace Ocelot.ServiceDiscovery
-{
-    public interface IServiceDiscoveryProvider
-    {
-         Task<List<Service>> Get();
-    }
-}
