+++ src/Ocelot/ServiceDiscovery/Providers/IServiceDiscoveryProvider.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using Ocelot.Values;
+
+namespace Ocelot.ServiceDiscovery.Providers
+{
+    public interface IServiceDiscoveryProvider
+    {
+         Task<List<Service>> Get();
+    }
+}
