--- src/Ocelot/ServiceDiscovery/UnableToFindServiceDiscoveryProviderError.cs
@@ -1,12 +1,0 @@
-using Ocelot.Errors;
-
-namespace Ocelot.ServiceDiscovery
-{
-    public class UnableToFindServiceDiscoveryProviderError : Error
-    {
-        public UnableToFindServiceDiscoveryProviderError(string message) 
-            : base(message, OcelotErrorCode.UnableToFindServiceDiscoveryProviderError)
-        {
-        }
-    }
-}
